+++ package.json
@@ -11,15 +11,14 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.3.0",
-    "sass": "^1.25.0"
+    "react-scripts": "^3.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "lint": "eslint src/**/*.js src/**/*.jsx .eslintrc.js"
+    "lint": "eslint src/**/*.js src/**/*.jsx .eslintrc.js src/**/*.tsx"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -37,6 +36,13 @@
     ]
   },
   "devDependencies": {
+    "@types/jest": "^24.9.0",
+    "@types/node": "^13.1.8",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-router-dom": "^5.1.3",
+    "@typescript-eslint/eslint-plugin": "^2.17.0",
+    "@typescript-eslint/parser": "^2.17.0",
     "babel-plugin-macros": "^2.8.0",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.0",
@@ -45,7 +51,9 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-react": "^7.18.0",
     "eslint-plugin-standard": "^4.0.1",
-    "react-git-info": "1.0.0"
+    "react-git-info": "1.0.0",
+    "sass": "^1.25.0",
+    "typescript": "^3.7.5"
   },
   "repository": {
     "type": "git",
